backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Armando Ezequiel Puerta <armando.ezequiel.puerta@gmail.com>"]
readme = "README.md"
license = "MIT License"
homepage = "https://github.com/EzequielPuerta/runiverse"
repository = "https://github.com/EzequielPuerta/runiverse"
documentation = "https://github.com/EzequielPuerta/runiverse"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.1.0"
django = "^5.2"
djangorestframework = "^3.16.0"
suitable-class-finder = "^0.1.0"
dnspython = "^2.7.0"
psycopg2-binary = "^2.9.10"
sqlalchemy = "^2.0.40"
marshmallow-sqlalchemy = "^1.4.2"
alembic = "^1.15.2"

[tool.poetry.group.dev.dependencies]
pytest-cov = "^6.1.1"
pre-commit = "^4.2.0"
mypy = "^1.15.0"
isort = "^6.0.1"
black = {version = "^25.1.0", allow-prereleases = true}
pytest = "^8.3.5"
pytest-django = "^4.11.1"
ruff = "^0.11.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 79

[tool.black]
line-length = 79
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.mypy]
follow_imports = "silent"
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
disallow_untyped_defs = true
ignore_missing_imports = true
exclude = [
    "backend/manage.py",
    "backend/alembic/versions/*",
]

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[tool.ruff]
line-length = 88
extend-ignore = ["E203"]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.ruff.lint.per-file-ignores]
"backend/alembic/versions/*.py" = ["ALL"]
